+++ eslint/templates/ruleFileTemplate.mjs
@@ -1,6 +1,10 @@
 export const ruleFileTemplate = {
   id: "rule-file",
-  body: "${IMPORTS}\n${SETUP}\n${RULE}",
+  body: `
+    {{IMPORTS}}
+    {{SETUP}}
+    {{RULE}}
+  `,
   slots: {
     IMPORTS: { type: "ImportDeclaration", minOccurs: 1, maxOccurs: 10 },
     SETUP: { type: ["TSTypeAliasDeclaration", "TSInterfaceDeclaration", "VariableDeclaration", "FunctionDeclaration"], named: /^(?!rule$)/, minOccurs: 0, maxOccurs: 10 },