+++ eslint/templates/schemaFileTemplate.mjs
@@ -1,6 +1,10 @@
 export const schemaFileTemplate = {
   id: "schema-file",
-  body: "${IMPORTS}\n${SETUP}\n${SCHEMA}",
+  body: `
+    {{IMPORTS}}
+    {{SETUP}}
+    {{SCHEMA}}
+  `,
   slots: {
     IMPORTS: { type: "ImportDeclaration", minOccurs: 1, maxOccurs: 5 },
     SETUP: { type: ["TSTypeAliasDeclaration", "TSInterfaceDeclaration", "VariableDeclaration", "FunctionDeclaration"], named: /^(?!.*Schema$)/, minOccurs: 0, maxOccurs: 30 },