+++ eslint/templates/typesModuleTemplate.mjs
@@ -1,6 +1,9 @@
 export const typesModuleTemplate = {
   id: "types-module",
-  body: "${IMPORTS}\n${TYPES}",
+  body: `
+    {{IMPORTS}}
+    {{TYPES}}
+  `,
   slots: {
     IMPORTS: { type: "ImportDeclaration", minOccurs: 0, maxOccurs: 5 },
     TYPES: { type: ["TSTypeAliasDeclaration", "TSInterfaceDeclaration"], minOccurs: 1, maxOccurs: 20 },